% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGrob.R
\name{updateGrob}
\alias{updateGrob}
\title{Updates the grob object}
\usage{
updateGrob(x, remove_key = TRUE)
}
\arguments{
\item{x}{A ggplot2 graph.}

\item{remove_key}{Boolean to whether to remove the key from the legend. True
to remove.}
}
\description{
Updates the grob object
}
\keyword{internal}
