% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeLegendColour.R
\name{changeLegendColour}
\alias{changeLegendColour}
\title{Customise the text colour of your legend.}
\usage{
changeLegendColour(x, remove_key = TRUE)
}
\arguments{
\item{x}{A ggplot2 graph.}

\item{remove_key}{Boolean to whether to remove the key from the legend. True
to remove. Defaults to True.}
}
\description{
Customise the text colour of your legend.
}
